fix javafx
