+++ src/main/java/MainDuMain.java
@@ -1,0 +1,5 @@
+public class MainDuMain {
+    public static void main(final String[] args) {
+        Main.main(args);
+    }
+}
